+++ onnx/version_converter/convert.h
@@ -740,6 +740,13 @@ class DefaultVersionConverter : public BaseVersionConverter {
         OpSetID(15), OpSetID(16)));
       registerAdapter(make_unique<CompatibleAdapter>("PRelu",
         OpSetID(15), OpSetID(16)));
+
+      /******** 16 -> 17 ********/
+      registerAdapter(make_unique<CompatibleAdapter>("Pad",
+        OpSetID(16), OpSetID(17)));
+      registerAdapter(make_unique<CompatibleAdapter>("Shape",
+        OpSetID(16), OpSetID(17)));
+
     }
 
     ModelProto convert_version(